requirements.txt
bitstring==3.1.5
Pillow==4.2.1
